Migrate routes
