--- _src_/constants/env.ts
@@ -1,1 +1,0 @@
-export const PROD_ENV = process.env.NEXT_PUBLIC_VERCEL_ENV === 'production';
