--- _src_/constants/libraries.ts
@@ -1,47 +1,0 @@
-export const FeatherIcons = {
-  name: 'Feather Icons',
-  urlPattern: 'iconophor.com/icons/feather/:version/:icon',
-  github: 'https://github.com/feathericons/feather',
-  homepage: 'https://feathericons.com',
-  unpkg: 'https://unpkg.com/browse/feather-icons/',
-};
-
-export const MaterialDesign = {
-  name: 'Material Design',
-  urlPattern: 'iconophor.com/icons/material/:version/:style/:icon',
-  github: 'https://github.com/material-icons/material-icons',
-  homepage: 'https://material.io/tools/icons',
-  unpkg: 'https://unpkg.com/browse/@material-icons/svg/',
-};
-
-export const FontAwesome = {
-  name: 'Font Awesome',
-  urlPattern: 'iconophor.com/icons/fontawesome/:version/:style/:icon',
-  github: 'https://github.com/FortAwesome/Font-Awesome',
-  homepage: 'https://fontawesome.com',
-  unpkg: 'https://unpkg.com/browse/@fortawesome/fontawesome-free/',
-};
-
-export const Heroicons = {
-  name: 'Heroicons',
-  urlPattern: 'iconophor.com/icons/heroicons/:version/:style/:icon',
-  github: 'https://github.com/tailwindlabs/heroicons',
-  homepage: 'https://heroicons.com',
-  unpkg: 'https://unpkg.com/browse/heroicons/',
-};
-
-export const SanityIcons = {
-  name: '@sanity/icons',
-  urlPattern: 'iconophor.com/icons/sanity-icons/:version/:icon',
-  github: 'https://github.com/sanity-io/design',
-  homepage: 'https://www.sanity.io/ui',
-  unpkg: 'https://unpkg.com/browse/@sanity/icons/',
-};
-
-export const libraries = [
-  FeatherIcons,
-  FontAwesome,
-  Heroicons,
-  MaterialDesign,
-  SanityIcons,
-];
