--- _src_/constants/svg-attributes.ts
@@ -1,16 +1,0 @@
-export const svgAttributes = [
-  'fill',
-  'stroke',
-  'width',
-  'height',
-  'color',
-  'fill-opacity',
-  'fill-rule',
-  'stroke-dasharray',
-  'stroke-dashoffset',
-  'stroke-linecap',
-  'stroke-linejoin',
-  'stroke-miterlimit',
-  'stroke-opacity',
-  'stroke-width',
-] as const;
