--- _src_/env.d.ts
@@ -1,1 +1,0 @@
-/// <reference types="astro/client" />
