--- _src_/layouts/Layout.astro
@@ -1,51 +1,0 @@
----
-import '@/styles/main.css';
-
-export interface Props {
-  title: string;
-  description: string;
-}
-
-const year = new Date().getFullYear();
-const title = Astro.props.title;
-const description = Astro.props.description;
-const image = 'https://www.iconophor.com/iconophor.png';
----
-
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width" />
-    <title>{title}</title>
-    <meta content={title} property="og:title" />
-    <meta content={description} name="description" />
-    <meta content={description} property="og:description" />
-    <meta content={image} property="og:image" />
-    <meta content={image} property="twitter:image" />
-    <meta content="https://www.iconophor.com" property="og:url" />
-    <link color="#101010" href="/favicon.svg" rel="icon" type="image/svg+xml" />
-    <link
-      color="#101010"
-      href="/favicon.svg"
-      rel="mask-icon"
-      type="image/svg+xml"
-    />
-  </head>
-  <body>
-    <main class="constrain-x mt-72 md:mt-144">
-      <slot />
-    </main>
-    <footer class="constrain-x mt-72 md:mt-144 mb-24">
-      <div class="flex flex-col md:flex-row gap-y-8 justify-between">
-        <p>© {year} Iconophor</p>
-        <p>
-          Made by
-          <a class="underline" href="https://www.msfragala.com">
-            Mitchell Fragala
-          </a>
-        </p>
-      </div>
-    </footer>
-  </body>
-</html>
