--- _src_/lib/exception.ts
@@ -1,10 +1,0 @@
-export class Exception extends Error {
-  status: number;
-  statusText: string;
-
-  constructor(status: number, message: string) {
-    super(message);
-    this.status = status;
-    this.statusText = message;
-  }
-}
