--- _src_/pages/icons/fontawesome/[version]/[style]/[icon].ts
@@ -1,9 +1,0 @@
-import type { APIRoute } from 'astro';
-import { catchError } from '../lib/catch-error';
-import { iconHandler } from '../lib/icon-handler';
-
-export const GET: APIRoute = catchError(
-  iconHandler(({ version, icon, style }) => {
-    return `https://unpkg.com/@fortawesome/fontawesome-free@${version}/svgs/${style}/${icon}.svg`;
-  })
-);
