--- _src_/pages/icons/heroicons/[version]/[style]/[icon].ts
@@ -1,9 +1,0 @@
-import type { APIRoute } from 'astro';
-import { catchError } from '../lib/catch-error';
-import { iconHandler } from '../lib/icon-handler';
-
-export const GET: APIRoute = catchError(
-  iconHandler(({ version, icon, style }) => {
-    return `https://raw.githubusercontent.com/tailwindlabs/heroicons/v${version}/optimized/24/${style}/${icon}.svg`;
-  })
-);
