--- _src_/pages/index.astro
@@ -1,100 +1,0 @@
----
-import { Playground } from '@/components/playground';
-import { libraries } from '@/constants/libraries';
-import { svgAttributes } from '@/constants/svg-attributes';
-import Layout from '@/layouts/Layout.astro';
----
-
-<Layout
-  title="Iconophor"
-  description="An open-source CDN for popular SVG icon libraries."
->
-  <div class="space-y-60" style={{ borderColor: 'red' }}>
-    <section>
-      <h1 class="heading-xl text-center">Iconophor</h1>
-      <p class="mt-48">
-        Iconophor is an
-        {' '}
-        <a class="link" href="https://github.com/msfragala/iconophor">
-          open-source
-        </a>
-        {' '}
-        CDN for popular SVG icon libraries that enables you to load icons as images
-        or using
-        {' '}
-        <a
-          class="link"
-          href="https://developer.mozilla.org/en-US/docs/Web/CSS/mask"
-          >CSS masks</a
-        >
-        {' '}
-        while still controlling the fill, stroke width, etc. via query parameters.
-        Embedding the SVG source code inline is a common way to use icons in UI frameworks,
-        but this can cause undesirable page bloat. Loading icons as images or masks
-        instead can be beneficial for performance in some contexts by reducing duplicate
-        markup and leveraging of caching.
-      </p>
-    </section>
-    <section>
-      <h2 class="heading-lg mb-24">Playground</h2>
-      <p class="mb-24">Try out the API in this demo with Feather Icons:</p>
-      <Playground client:load />
-    </section>
-    <section>
-      <h2 class="heading-lg mb-24">Icon libraries</h2>
-      <ul class="grid grid-cols-1 gap-y-36 w-full">
-        {
-          libraries.map((library) => (
-            <li>
-              <h3 class="mb-4">
-                <a
-                  class="heading-md text-pink-300 hover:underline focus-visible:underline"
-                  href={library.homepage}
-                >
-                  {library.name}
-                </a>
-              </h3>
-              <p class="mb-8">
-                <a
-                  aria-label={`${library.name} on GitHub`}
-                  class="text-sm hover:underline focus-visible:underline"
-                  href={library.github}
-                >
-                  GitHub
-                </a>{' '}
-                •{' '}
-                <a
-                  aria-label={`${library.name} on Unpkg`}
-                  class="text-sm hover:underline focus-visible:underline"
-                  href={library.unpkg}
-                >
-                  Unpkg
-                </a>
-              </p>
-              <p class="pill break-words max-w-full overflow-hidden">
-                {library.urlPattern}
-              </p>
-            </li>
-          ))
-        }
-      </ul>
-    </section>
-    <section>
-      <h2 class="heading-lg mb-24">Query parameters</h2>
-      <ul class="space-y-12 sm:columns-2 sm:space-y-16">
-        {
-          svgAttributes.map((name) => (
-            <li>
-              <a
-                class="pill"
-                href={`https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/${name}`}
-              >
-                <code>?{name}</code>
-              </a>
-            </li>
-          ))
-        }
-      </ul>
-    </section>
-  </div>
-</Layout>
