+++ package.json
@@ -17,6 +17,7 @@
 		"tabWidth": 2,
 		"useTabs": true,
 		"quoteProps": "consistent",
+		"printWidth": 100,
 		"singleQuote": true,
 		"plugins": [
 			"prettier-plugin-svelte"
@@ -50,6 +51,7 @@
 	"devDependencies": {
 		"@sveltejs/adapter-auto": "^2.0.0",
 		"@sveltejs/kit": "^1.20.4",
+		"@types/svg-parser": "^2.0.4",
 		"@typescript-eslint/eslint-plugin": "^6.0.0",
 		"@typescript-eslint/parser": "^6.7.5",
 		"autoprefixer": "^10.4.14",
