+++ src/app.css
@@ -1,3 +1,107 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
+
+@layer base {
+	:root {
+		--colors-background: #101010;
+		--colors-background-soft: #242424;
+		--colors-text: #ffffff;
+		--colors-text-soft: #adadad;
+		--colors-border: #333;
+		--colors-focus-ring: theme('colors.blue.500');
+		--focus-ring-offset: 0.3em;
+		--focus-ring-width: 2px;
+	}
+
+	* {
+		margin: 0;
+		padding: 0;
+		box-sizing: border-box;
+	}
+
+	html,
+	body {
+		background: var(--colors-background);
+		color: var(--colors-text);
+		font-size: 16px;
+	}
+
+	body {
+		max-width: 100vw;
+		overflow-x: clip;
+	}
+
+	*:focus:not(:focus-visible) {
+		outline: none !important;
+	}
+
+	*:focus-visible {
+		isolation: isolate;
+		outline: var(--focus-ring-width) solid var(--colors-focus-ring);
+		outline-offset: var(--focus-ring-offset);
+		z-index: 1;
+	}
+
+	:is(h1, h2, h3, h4, p) a:focus-visible {
+		display: inline-block;
+		background: var(--colors-background);
+	}
+
+	button:focus-visible,
+	input:focus-visible,
+	select:focus-visible,
+	textarea:focus-visible {
+		--focus-ring-offset: 0;
+	}
+}
+
+.heading-xl {
+	@apply text-36 xs:text-60 lg:text-72 tracking-wide font-bold;
+}
+
+.heading-lg {
+	@apply text-lg tracking-wide font-bold;
+}
+
+.heading-md {
+	@apply tracking-wide font-medium;
+	font-size: 18px;
+}
+
+.pill {
+	@apply bg-background-soft rounded px-12 py-8 font-mono block w-content text-sm;
+	--focus-ring-offset: 0;
+}
+
+.inline-code {
+	@apply pill;
+	display: inline;
+	padding: 0.25em;
+}
+
+.input {
+	@apply bg-background-soft rounded px-12 py-8;
+}
+
+.link {
+	@apply text-pink-300 hover:underline focus-visible:underline;
+}
+
+.constrain-x {
+	@apply max-w-640 w-full mx-auto px-24 xs:px-36 md:px-0;
+}
+
+@layer utilities {
+	.columns-1 {
+		column-count: 1;
+	}
+
+	.columns-2 {
+		column-count: 2;
+	}
+
+	.mask-svg {
+		mask: var(--mask-svg) center / contain no-repeat;
+	}
+}
