+++ src/routes/icons/material/[version]/[style]/[icon]/+server.ts
@@ -1,0 +1,9 @@
+import type { RequestHandler } from '@sveltejs/kit';
+import { iconHandler } from '@/lib/icon-handler';
+
+export const GET: RequestHandler = iconHandler(
+	['version', 'icon', 'style'],
+	({ version, icon, style }) => {
+		return `https://unpkg.com/@material-icons/svg@${version}/svg/${icon}/${style}.svg`;
+	}
+);
